+++ lib/flock.dart
@@ -6,6 +6,7 @@ export 'base/create_store.dart';
 export 'base/types.dart';
 export 'contexts/use_case.dart';
 export 'enhancers/batch_subscribe.dart';
+export 'enhancers/only_publish_when.dart';
 export 'enhancers/print_events.dart';
 export 'enhancers/with_side_effect.dart';
 export 'integration/project_to_stream.dart';